pyproject.toml
[project]
name = "healthcare-pulmonary-diagnosis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dvc>=3.66.1",
    "ensure>=1.0.4",
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "gdown>=5.2.1",
    "joblib>=1.5.3",
    "kagglehub>=1.0.0",
    "matplotlib>=3.10.8",
    "mlflow>=3.10.0",
    "numpy>=2.4.2",
    "pandas>=2.3.3",
    "pillow>=12.1.1",
    "pymongo[srv]==3.12",
    "python-box>=7.3.3",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.0",
    "seaborn>=0.13.2",
    "tensorflow>=2.20.0",
    "tqdm>=4.67.3",
    "types-pyyaml>=6.0.12.20250915",
    "umap-learn>=0.5.11",
    "xgboost>=3.2.0",
]
